chore: hoist asset
