+++ docs/reference/asset.md
@@ -1,0 +1,1 @@
+::: kolena.asset
