+++ mkdocs.yml
@@ -43,6 +43,7 @@ nav:
   - API Reference:
       - API Reference: reference/index.md
       - <code>kolena.annotation</code>: reference/annotation.md
+      - <code>kolena.asset</code>: reference/asset.md
       - <code>kolena.initialize</code>: reference/initialize.md
       - <code>kolena.errors</code>: reference/errors.md
       - <code>kolena.dataset</code>:
