+++ tests/unit/workflow/test_asset.py
@@ -19,7 +19,7 @@ import pydantic
 import pytest
 
 from kolena._utils.datatypes import DATA_TYPE_FIELD
-from kolena.workflow.asset import _AssetType
+from kolena.asset import _AssetType
 from kolena.workflow.asset import Asset
 from kolena.workflow.asset import BinaryAsset
 from kolena.workflow.asset import ImageAsset
